refactor: centralize tool-use snapshot schemas
